+++ src/Comparison/ComparisonStrategy.ts
@@ -6,5 +6,32 @@ import { EQUALS } from "./Equals"
 import { CONTAINS_ALL_WORDS } from "./ContainsAllWords"
 import { BITAP } from "./Bitap"
 import { REGULAR_EXPRESSION } from "./RegularExpression"
+import { COSINE } from "./Cosine"
+import { EUCLIDEAN } from "./Euclidean"
+import { HAMMING } from "./Hamming"
+import { JACCARD } from "./Jaccard"
+import { JARO_WINKLER } from "./JaroWinkler"
+import { LEVENSHTEIN } from "./Levenshtein"
+import { LONGEST_COMMON_SUBSTRING } from "./LongestCommonSubstring"
+import { QGRAM } from "./qgram"
+import { DAMERAU } from "./Damerau"
 
-export { FUZZY_SEQUENCE, STARTS_WITH, ENDS_WITH, CONTAINS, EQUALS, CONTAINS_ALL_WORDS, BITAP, REGULAR_EXPRESSION }
+export {
+	FUZZY_SEQUENCE,
+	STARTS_WITH,
+	ENDS_WITH,
+	CONTAINS,
+	EQUALS,
+	CONTAINS_ALL_WORDS,
+	BITAP,
+	REGULAR_EXPRESSION,
+	COSINE,
+	EUCLIDEAN,
+	HAMMING,
+	JACCARD,
+	JARO_WINKLER,
+	LEVENSHTEIN,
+	LONGEST_COMMON_SUBSTRING,
+	QGRAM,
+	DAMERAU
+}